+++ src/Frames/ColumnTypeable.hs
@@ -38,7 +38,8 @@ instance Parseable Bool where
 instance Parseable Int where
 instance Parseable Float where
 instance Parseable Double where
-  -- Some CSV's export Doubles in a format like '1,000.00', filtering out commas lets us parse those sucessfully
+  -- Some CSV's export Doubles in a format like '1,000.00', filtering
+  -- out commas lets us parse those sucessfully
   parse = fmap Definitely . fromText . T.filter (/= ',')
 instance Parseable T.Text where